ImageFileSystem 제거 및 FileMediaTypeResolver 생성
