+++ src/main/java/com/jydev/noticeboard/file/FileSystem.java
@@ -2,6 +2,7 @@ package com.jydev.noticeboard.file;
 
 import com.jydev.noticeboard.file.model.FileType;
 import com.jydev.noticeboard.file.model.StoreFile;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
 import org.springframework.http.MediaType;
@@ -10,10 +11,14 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Optional;
 import java.util.UUID;
 
+@RequiredArgsConstructor
 @Component
 public class FileSystem {
+
+    private final FileMediaTypeResolver fileMediaTypeResolver;
 
     @Value("${file.dir}")
     protected String fileDir;
@@ -23,6 +28,11 @@ public class FileSystem {
 
     private String getFullFilePath(String filePath){
         return fileDir+filePath;
+    }
+
+    public static String getExt(String fileName) {
+        int extIndex = fileName.lastIndexOf(".");
+        return fileName.substring(extIndex+1);
     }
 
     public StoreFile storeFile(MultipartFile file, FileType fileType) throws IOException {
@@ -37,17 +47,16 @@ public class FileSystem {
         return UUID.randomUUID()+"."+ getExt(fileName);
     }
 
-    public String getExt(String fileName) {
-        int extIndex = fileName.lastIndexOf(".");
-        return fileName.substring(extIndex+1);
-    }
-
     public String getFileProtocolUrl(String fileName){
         return "file:"+getFullFilePath(fileName);
     }
 
     public String getHttpProtocolUrl(String fileName,FileType fileType){
         return url+"file/"+fileType.getProtocol()+"/"+fileName;
+    }
+
+    public Optional<MediaType> getMediaType(FileType fileType , String fileName){
+        return fileMediaTypeResolver.resolve(fileType,fileName);
     }
