+++ src/test/java/com/jydev/noticeboard/file/util/FileDependency.java
@@ -1,9 +1,8 @@
 package com.jydev.noticeboard.file.util;
 
+import com.jydev.noticeboard.file.FileMediaTypeResolver;
 import com.jydev.noticeboard.file.FileSystem;
-import com.jydev.noticeboard.file.ImageFileSystem;
 
 public class FileDependency {
-    public final static FileSystem fileSystem = new FileSystem();
-    public final static ImageFileSystem imageFileSystem = new ImageFileSystem();
+    public final static FileSystem fileSystem = new FileSystem(new FileMediaTypeResolver());
 }
