+++ src/test/java/com/jydev/noticeboard/user/util/UserDependency.java
@@ -1,6 +1,5 @@
 package com.jydev.noticeboard.user.util;
 
-import com.jydev.noticeboard.file.ImageFileSystem;
 import com.jydev.noticeboard.file.util.FileDependency;
 import com.jydev.noticeboard.user.model.Mapper.UserMapper;
 import com.jydev.noticeboard.user.repository.LoginRepository;
@@ -11,7 +10,7 @@ import com.jydev.noticeboard.user.service.UserService;
 import com.jydev.noticeboard.user.service.UserServiceImpl;
 
 public class UserDependency {
-    public static final UserMapper userMapper = new UserMapper(FileDependency.imageFileSystem);
+    public static final UserMapper userMapper = new UserMapper(FileDependency.fileSystem);
 
     public static UserRepository getUserRepository(){
         return new MemoryUserRepository(userMapper);
